# Makes IonValue.clone iterative and improves clone performance.