--- a/src/com/amazon/ion/impl/lite/IonClobLite.java
+++ b/src/com/amazon/ion/impl/lite/IonClobLite.java
@@ -48,7 +48,7 @@ final class IonClobLite
     }
 
     @Override
-    IonClobLite clone(IonContext context)
+    IonValueLite shallowClone(IonContext context)
     {
         return new IonClobLite(this, context);
     }
@@ -56,7 +56,7 @@ final class IonClobLite
     @Override
     public IonClobLite clone()
     {
-        return clone(ContainerlessContext.wrap(getSystem()));
+        return (IonClobLite) shallowClone(ContainerlessContext.wrap(getSystem()));
     }
 
     @Override