… src/com/amazon/ion/impl/lite/IonValueLite.java
… test/com/amazon/ion/CloneTest.java
… test/com/amazon/ion/IonValueTest.java