*Description of changes:*

This PR introduces the concept of "shallow clone" vs. "deep clone". Values of every type have a "shallow clone" implementation. For scalars, shallow clone clones all information about the value. For containers, shallow clone clones the container but not its children. Deep clone is only applicable for containers, and clones the container and all child values. `IonContainerLite.deepClone` contains the bulk of the iterative clone logic added in this PR; this was previously performed recursively (via `IonValue.clone`) in `IonContainerLite`'s copy constructor.

A key goal of this PR was to be at least performance-neutral. In order to achieve that, I had to make some optimizations that were not required to simply get the new implementation to work. This involved:
* Avoiding `SymbolToken` allocations where possible.
* Avoiding calling field name and annotation retrieval methods repetitively when direct access was possible, and
* Lazily instantiating IonStructLite's field map, which is a map from name to array index that is created to optimize field lookups when the struct grows to have more than 5 fields. Previously, this was eagerly copied during clone, contributing to the bulk of the time spent when cloning large structs. Performance tests (below) indicated that creating the field map lazily performed better *even when the cloned value was traversed and modified, requiring the field map to be materialized later*.

I also optimized `IonStructLite.doClone`, which provides the implementation for the public APIs `IonStruct.cloneAndRetain` and `IonStruct.cloneAndRemove`. The key optimizations here were:
* Avoiding allocating new `SymbolToken`s for annotations.
* Allocating the new `_children` array once rather than potentially going through multiple growth cycles by calling `IonStructLite.add`. Avoiding `add` also skips a lot of repetitive validation.
* Transitive improvements to `IonValue.clone()`, as described above and quantified below.

Some of the small optimizations described above also seemed to have a small positive effect on `IonValue.load`, quantified below. Further, I believe that there is additional performance improvement potential in the IonValue DOM, outside the scope of this PR.

### Performance

Datasets:

* entry.10n - 40 KB, containing deeply-nested structs with recurring structure, short- and medium-length strings, and numeric values.
* committed.10n - 140 B, single top-level struct with local symbol table.
* log.10n - 22MB, containing nested structs with recurring structure, symbols, annotations, and numeric values.
* bigInts10MB.10n - 10 MB of top-level integers that are larger than a Java long.

#### IonValue.clone

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| log.10n                 | 262043            | 155021 (-41%) |
| entry.10n              |  248                 |  178 (-28%)    |
| committed.10n    |  1.534               | 1.435 (-6%)   |
| bigInts10MB.10n  | 27050              | 19963 (-26%) |

#### IonValue.clone, then traverse and modify structs, forcing struct field maps to be materialized

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| log.10n                 | 614340            | 391448 (-36%) |
| entry.10n              |  270                 |  189 (-30%)    |
| committed.10n    |  1.557               | 1.459 (-6%)   |
| bigInts10MB.10n  | 34116              | 26237 (-23%) |

#### IonValue.load

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| log.10n                 | 1141580            | 1091150 (-4%) |
| entry.10n              |  1555                 |  1295 (-17%)    |
| committed.10n    |  5.132               | 5.095 (-1%)   |
| bigInts10MB.10n  | 127466              | 126964 (~0%) |

#### IonStruct.cloneAndRemove, removing the largest child value from log.ion, leaving only scalar children

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| log.10n                 | 209791            | 50054 (-76%) |

#### IonStruct.cloneAndRetain, retaining only the largest child value from log.ion, removing all scalar children

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| log.10n                 | 299065            | 198688 (-34%) |

#### IonStruct.cloneAndRetain, retaining only the largest child value from log.ion, removing all scalar children, then traverse and modify up to depth 3, forcing struct field maps to be materialized

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| log.10n                 | 656454            | 384255 (-41%) |

#### IonStruct.cloneAndRemove, removing a large blob value from events.10n and retaining annotated fields

`events.10n` is a 7.5MB binary Ion file containing about 2000 structs with annotated field names. The bulk of each struct is a large blob.

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| events.10n                 | 978            | 518 (-47%) |

#### IonStruct.cloneAndRetain, retaining only two annotated fields

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| events.10n                 | 968            | 549 (-43%) |


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.